+++ src/model/queries/SubGraphQuery.java
@@ -24,16 +24,16 @@ public class SubGraphQuery extends GraphQuery {
 
     @Override
     public Object executeQueryOnSummary() {
-        Float mergedWeight = 0f;
+        Integer mergedWeight = 0;
 
         for (GraphSketch sketch : this.graphSummary.getGraphSketches()) {
-            Float currentWeight = 0f;
+            Integer currentWeight = 0;
             for (Pair<String> pair : this.subGraph) {
 
                 int hashedA = (int)sketch.getHash().hashToBin(pair.getA());
                 int hashedB = (int)sketch.getHash().hashToBin(pair.getB());
 
-                Float weight = sketch.getAdjMatrix()[hashedA][hashedB];
+                Integer weight = sketch.getAdjMatrix()[hashedA][hashedB];
 
                 if (weight != null) {
                     currentWeight += weight;
@@ -50,7 +50,7 @@ public class SubGraphQuery extends GraphQuery {
 
     @Override
     public Object executeQueryOnOriginal() {
-        Float weight = 0f;
+        Integer weight = 0;
 
         for (Pair<String> pair : this.subGraph) {
 
@@ -89,8 +89,8 @@ public class SubGraphQuery extends GraphQuery {
             }
 
             GraphQuery testQuery = new SubGraphQuery(graphSummary, subGraph);
-            Float summaryResult = (Float)testQuery.executeQueryOnSummary();
-            Float originalResult = (Float)testQuery.executeQueryOnOriginal();
+            Integer summaryResult = (Integer)testQuery.executeQueryOnSummary();
+            Integer originalResult = (Integer)testQuery.executeQueryOnOriginal();
             //System.out.println(String.format("Summary: %.4f; Original: %.4f", summaryResult, originalResult));
             if (GraphQuery.assertAquality(summaryResult, originalResult)) {
                 NrOfCorrectQueries++;
